Extra game functionality.
Added Lives, Point, and InGame variable and a StartGame variable.

Added "Press Enter" graphics to start the game.

Draws the points in the top left of the screen, draws your lives below the points counter, lives are represented as the ship texture, like in the original game.

Playing field now handles bullet collision logic, it also handles when the player collides with an asteroid, such as decrementing the Lives counter, or Game Over if the lives are Zero.